res made the following changes:   - Removed the now-unneeded scfstrset.cpp and path.cpp files.   - Tweaked the driver DB to have the GeForce rules not match the     beginning of the GL renderer string (on Macs, this string starts     with "NVIDIA GeForce...").   - Converted sprcal3d to new SCF and made some smaller tweaks.   - Fixed some VC8 warnings.   - Fixed a subtle bug in csRenderBuffer that caused the last lock     type to be stored incorrectly.   - Tweaked csFrameDataHolder<> so it now releases memory after a     while as well.
git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/trunk@22966 8cc4aa7f-3514-0410-904f-f2cc9021211c

+++ plugins/mesh/genmesh/object/genmesh.h
@@ -530,7 +530,7 @@ private:
   csVector2* mesh_texels;
   csVector3* mesh_normals;
   csColor4* mesh_colors;
-  int num_mesh_vertices;
+  size_t num_mesh_vertices;
   csVector3* mesh_tri_normals;
 
   bool autonormals;
@@ -592,11 +592,11 @@ private:
    * Compress vertices. This is for CalculateNormals().
    */
   bool CompressVertices (
-	csVector3* orig_verts, int orig_num_vts,
-	csVector3*& new_verts, int& new_num_vts,
-	csTriangle* orig_tris, int num_tris,
+	csVector3* orig_verts, size_t orig_num_vts,
+	csVector3*& new_verts, size_t& new_num_vts,
+	csTriangle* orig_tris, size_t num_tris,
 	csTriangle*& new_tris,
-	int*& mapping);
+	size_t*& mapping);
 
   /**
    * Setup this factory. This function will check if setup is needed.
@@ -642,7 +642,7 @@ public:
   }
   iMaterialWrapper* GetMaterialWrapper () const { return material; }
   void SetVertexCount (int n);
-  int GetVertexCount () const { return num_mesh_vertices; }
+  int GetVertexCount () const { return (int)num_mesh_vertices; }
   csVector3* GetVertices ()
   {
     SetupFactory ();
